fix for 10gb transfer

+++ src/main/java/org/onedatashare/server/module/http/HttpResource.java
@@ -294,7 +294,7 @@ public class HttpResource extends Resource<HttpSession, HttpResource> {
          */
         public Flux<Slice> tap(long sliceSize) {
             int sliceSizeInt = Math.toIntExact(sliceSize);
-            int sizeInt = Math.toIntExact(size);
+            long sizeInt = size;
             InputStream inputStream = null;
             try {
                 inputStream = fileContent.getInputStream();
@@ -316,7 +316,7 @@ public class HttpResource extends Resource<HttpSession, HttpResource> {
                             }
                             sink.next(new Slice(b));
                         } else {
-                            int remaining = sizeInt - state;
+                            int remaining =  Math.toIntExact(sizeInt - state);
                             byte[] b = new byte[remaining];
                             try {
                                 // Fix for buggy PDF files - Else the PDF files are corrupted
