+++ src/main/java/org/onedatashare/server/module/vfs/VfsResource.java
@@ -198,7 +198,7 @@ public class VfsResource extends Resource<VfsSession, VfsResource> {
 
         public Flux<Slice> tap(long sliceSize) {
             int sliceSizeInt = Math.toIntExact(sliceSize);
-            int sizeInt = Math.toIntExact(size);
+            long sizeInt = size;
             InputStream inputStream = null;
             try {
                 inputStream = fileContent.getInputStream();
@@ -220,7 +220,7 @@ public class VfsResource extends Resource<VfsSession, VfsResource> {
                             }
                             sink.next(new Slice(b));
                         } else {
-                            int remaining = sizeInt - state;
+                            int remaining = Math.toIntExact(sizeInt - state);
                             byte[] b = new byte[remaining];
                             try {
                                 // Fix for buggy PDF files - Else the PDF files are corrupted
